+++ crates/core/config/src/lib.rs
@@ -1,4 +1,6 @@
+mod available_models;
 mod config;
 mod parse_replace_pairs;
 
+pub use available_models::{get_default_model, AVAILABLE_MODELS};
 pub use config::{Config, ShortcutConfigProvider};